projects/youtube-live-insights/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
attrs==22.1.0
aws-cdk-constructs==2.31.3
aws-cdk-lib==2.123.0
aws-cdk.asset-awscli-v1==2.2.202
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v6==2.0.3
aws-cdk.aws-neptune-alpha==2.123.0a0
AWSIoTPythonSDK==1.5.4
awslambdaric==2.0.11
b-aws-websocket-api==2.0.0
b-stage-deployment==1.0.1
boto3==1.26.59
botocore==1.29.165
cattrs==23.1.2
certifi==2024.7.4
charset-normalizer==3.3.2
chat-downloader==0.2.8
click==8.1.3
colorlog==6.8.2
constructs==10.3.0
coverage==7.2.3
diagrams==0.23.4
distro==1.9.0
docstring_parser==0.16
exceptiongroup==1.2.1
graphviz==0.20.3
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
importlib_resources==6.4.0
iniconfig==2.0.0
inquirerpy==0.3.4
isodate==0.6.1
Jinja2==3.1.4
jmespath==1.0.1
jsii==1.101.0
lambda-forge==2.1.10
MarkupSafe==2.1.5
openai==1.35.10
packaging==24.1
pfzy==0.3.4
pluggy==1.5.0
prompt_toolkit==3.0.47
publication==0.0.3
py==1.11.0
pydantic==2.7.3
pydantic_core==2.18.4
pyfiglet==1.0.2
pytest==6.2.5
pytest-cov==2.12.1
pytest-sugar==1.0.0
pytest-timeout==1.4.2
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytube==15.0.0
requests==2.31.0
s3transfer==0.6.2
simplejson==3.18.4
six==1.16.0
sniffio==1.3.1
tabulate==0.9.0
termcolor==2.4.0
toml==0.10.2
tqdm==4.66.4
typed-ast==1.5.5
typeguard==2.13.3
typing_extensions==4.12.2
urllib3==1.26.19
wcwidth==0.2.13
websocket-client==1.8.0
websockets==11.0.3
zipp==3.19.2
